round returns int by default

+++ homeassistant/components/netatmo/light.py
@@ -174,7 +174,7 @@ class NetatmoLight(NetatmoModuleEntity, LightEntity):
         """Turn light on."""
         if ATTR_BRIGHTNESS in kwargs:
             await self.device.async_set_brightness(
-                int(round((kwargs[ATTR_BRIGHTNESS]) / 2.55))
+                round((kwargs[ATTR_BRIGHTNESS]) / 2.55)
             )
 
         else:
@@ -196,6 +196,6 @@ class NetatmoLight(NetatmoModuleEntity, LightEntity):
 
         if (brightness := self.device.brightness) is not None:
             # Netatmo uses a range of [0, 100] to control brightness
-            self._attr_brightness = int(round(brightness * 2.55))
+            self._attr_brightness = round(brightness * 2.55)
         else:
             self._attr_brightness = None
